test: add detailed acceptance tests for payment groups and restructure workflows
- Introduced `PaymentDetailCest` for comprehensive testing of payment group behaviors, URLs, email actions, and grid rendering.
- Decoupled shared payment logic into `PaymentAcceptanceCest` for better abstraction and reusability.
- Refactored `PaymentCest` to extend `PaymentAcceptanceCest` and relocated specific tests.
- Updated GitHub workflows to add dedicated shards for `PaymentDetailCest` and `PaymentDashboardCest`.
